feat: merge model files for SD client
