--- src/main/kotlin/net/atos/client/sd/model/templates/User.kt
@@ -1,13 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2024 Lifely
- * SPDX-License-Identifier: EUPL-1.2+
- */
-package net.atos.client.sd.model.templates
-
-import nl.lifely.zac.util.NoArgConstructor
-
-@NoArgConstructor
-class User {
-    lateinit var id: String
-    lateinit var name: String
-}
