--- src/main/kotlin/net/atos/client/sd/model/wizard/Document.kt
@@ -1,7 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2024 Lifely
- * SPDX-License-Identifier: EUPL-1.2+
- */
-package net.atos.client.sd.model.wizard
-
-class Document
