--- src/main/kotlin/net/atos/client/sd/model/wizard/File.kt
@@ -1,10 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2024 Lifely
- * SPDX-License-Identifier: EUPL-1.2+
- */
-package net.atos.client.sd.model.wizard
-
-class File {
-    lateinit var fileName: String
-    lateinit var document: Document
-}
