--- src/main/kotlin/net/atos/client/sd/model/wizard/Selection.kt
@@ -1,15 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2024 Lifely
- * SPDX-License-Identifier: EUPL-1.2+
- */
-package net.atos.client.sd.model.wizard
-
-import jakarta.json.bind.annotation.JsonbProperty
-
-class Selection {
-    @JsonbProperty("TemplateGroup")
-    var templateGroup: String? = null
-
-    @JsonbProperty("Template")
-    var template: String? = null
-}
