--- src/main/kotlin/net/atos/client/sd/model/wizard/SmartDocument.kt
@@ -1,12 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2024 Lifely
- * SPDX-License-Identifier: EUPL-1.2+
- */
-package net.atos.client.sd.model.wizard
-
-import jakarta.json.bind.annotation.JsonbProperty
-
-class SmartDocument {
-    @JsonbProperty("Selection")
-    lateinit var selection: Selection
-}
