+++ src/main/kotlin/net/atos/client/sd/model/wizard/UnattendedResponse.kt
@@ -10,3 +10,10 @@ class UnattendedResponse {
     @JsonbProperty("file")
     var files: List<File>? = null
 }
+
+class File {
+    lateinit var fileName: String
+    lateinit var document: Document
+}
+
+class Document
